+++ src/Umbraco.Tests/Web/UmbracoHelperTests.cs
@@ -1,183 +1,29 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Text;
 using Moq;
 using NUnit.Framework;
 using Umbraco.Core;
-using Umbraco.Tests.TestHelpers;
 using Umbraco.Core.Cache;
 using Umbraco.Core.Composing;
 using Umbraco.Core.Configuration;
 using Umbraco.Core.Logging;
+using Umbraco.Tests.TestHelpers;
 using Umbraco.Web;
 
-namespace Umbraco.Tests.FrontEnd
+namespace Umbraco.Tests.Web
 {
     [TestFixture]
     public class UmbracoHelperTests
-    {
-        private const string SampleWithAnchorElement = "Hello world, this is some text <a href='blah'>with a link</a>";
-        private const string SampleWithBoldAndAnchorElements = "Hello world, <b>this</b> is some text <a href='blah'>with a link</a>";
+    {   
 
         [TearDown]
         public void TearDown()
         {
             Current.Reset();
         }
-        [Test]
-        public static void Truncate_Simple()
-        {
-            var helper = new UmbracoHelper();
+        
 
-            var result = helper.Truncate(SampleWithAnchorElement, 25).ToString();
-
-            Assert.AreEqual("Hello world, this is some&hellip;", result);
-        }
-
-        [Test]
-        public static void When_Truncating_A_String_Ends_With_A_Space_We_Should_Trim_The_Space_Before_Appending_The_Ellipsis()
-        {
-            var helper = new UmbracoHelper();
-
-            var result = helper.Truncate(SampleWithAnchorElement, 26).ToString();
-
-            Assert.AreEqual("Hello world, this is some&hellip;", result);
-        }
-
-        [Test]
-        public static void Truncate_Inside_Word()
-        {
-            var helper = new UmbracoHelper();
-
-            var result = helper.Truncate(SampleWithAnchorElement, 24).ToString();
-
-            Assert.AreEqual("Hello world, this is som&hellip;", result);
-        }
-
-        [Test]
-        public static void Truncate_With_Tag()
-        {
-            var helper = new UmbracoHelper();
-
-            var result = helper.Truncate(SampleWithAnchorElement, 35).ToString();
-
-            Assert.AreEqual("Hello world, this is some text <a href='blah'>with&hellip;</a>", result);
-        }
-
-        [Test]
-        public static void Create_Encrypted_RouteString_From_Anonymous_Object()
-        {
-            var additionalRouteValues = new
-            {
-                key1 = "value1",
-                key2 = "value2",
-                Key3 = "Value3",
-                keY4 = "valuE4"
-            };
-
-            var encryptedRouteString = UmbracoHelper.CreateEncryptedRouteString(
-                "FormController",
-                "FormAction",
-                "",
-                additionalRouteValues
-                );
-
-            var result = encryptedRouteString.DecryptWithMachineKey();
-
-            const string expectedResult = "c=FormController&a=FormAction&ar=&key1=value1&key2=value2&Key3=Value3&keY4=valuE4";
-
-            Assert.AreEqual(expectedResult, result);
-        }
-
-        [Test]
-        public static void Create_Encrypted_RouteString_From_Dictionary()
-        {
-            var additionalRouteValues = new Dictionary<string, object>()
-            {
-                {"key1", "value1"},
-                {"key2", "value2"},
-                {"Key3", "Value3"},
-                {"keY4", "valuE4"}
-            };
-
-            var encryptedRouteString = UmbracoHelper.CreateEncryptedRouteString(
-                "FormController",
-                "FormAction",
-                "",
-                additionalRouteValues
-                );
-
-            var result = encryptedRouteString.DecryptWithMachineKey();
-
-            const string expectedResult = "c=FormController&a=FormAction&ar=&key1=value1&key2=value2&Key3=Value3&keY4=valuE4";
-
-            Assert.AreEqual(expectedResult, result);
-        }
-
-        [Test]
-        public static void Truncate_By_Words()
-        {
-            var helper = new UmbracoHelper();
-
-            var result = helper.TruncateByWords(SampleWithAnchorElement, 4).ToString();
-
-            Assert.AreEqual("Hello world, this is&hellip;", result);
-        }
-
-        [Test]
-        public static void Truncate_By_Words_With_Tag()
-        {
-            var helper = new UmbracoHelper();
-
-            var result = helper.TruncateByWords(SampleWithBoldAndAnchorElements, 4).ToString();
-
-            Assert.AreEqual("Hello world, <b>this</b> is&hellip;", result);
-        }
-
-        [Test]
-        public static void Truncate_By_Words_Mid_Tag()
-        {
-            var helper = new UmbracoHelper();
-
-            var result = helper.TruncateByWords(SampleWithAnchorElement, 7).ToString();
-
-            Assert.AreEqual("Hello world, this is some text <a href='blah'>with&hellip;</a>", result);
-        }
-
-        [Test]
-        public static void Strip_All_Html()
-        {
-            var helper = new UmbracoHelper();
-
-            var result = helper.StripHtml(SampleWithBoldAndAnchorElements, null).ToString();
-
-            Assert.AreEqual("Hello world, this is some text with a link", result);
-        }
-
-        [Test]
-        public static void Strip_Specific_Html()
-        {
-            string[] tags = { "b" };
-
-            var helper = new UmbracoHelper();
-
-            var result = helper.StripHtml(SampleWithBoldAndAnchorElements, tags).ToString();
-
-            Assert.AreEqual(SampleWithAnchorElement, result);
-        }
-
-        [Test]
-        public static void Strip_Invalid_Html()
-        {
-            const string text = "Hello world, <bthis</b> is some text <a href='blah'>with a link</a>";
-
-            var helper = new UmbracoHelper();
-
-            var result = helper.StripHtml(text).ToString();
-
-            Assert.AreEqual("Hello world, is some text with a link", result);
-        }
-
+        
         // ------- Int32 conversion tests
         [Test]
         public static void Converting_Boxed_34_To_An_Int_Returns_34()
