+++ src/Umbraco.Web/Editors/EntityController.cs
@@ -213,7 +213,7 @@ namespace Umbraco.Web.Editors
 
             if (type == UmbracoEntityTypes.Document)
             {
-                var foundUrl = Umbraco.Url(id);
+                var foundUrl = UmbracoContext.Url(id);
                 if (string.IsNullOrEmpty(foundUrl) == false && foundUrl != "#")
                 {
                     returnUrl = foundUrl;
