+++ src/Umbraco.Web/UmbracoHelper.cs
@@ -22,14 +22,11 @@ namespace Umbraco.Web
     /// </remarks>
     public class UmbracoHelper
     {
-        private static readonly HtmlStringUtilities StringUtilities = new HtmlStringUtilities();
-
         private readonly IPublishedContentQuery _publishedContentQuery;
         private readonly ITagQuery _tagQuery;
         private readonly MembershipHelper _membershipHelper;
         private readonly IUmbracoComponentRenderer _componentRenderer;
         private readonly ICultureDictionaryFactory _cultureDictionaryFactory;
-        private readonly UmbracoContext _umbracoContext;
 
         private IPublishedContent _currentPage;
         private ICultureDictionary _cultureDictionary;
@@ -39,28 +36,26 @@ namespace Umbraco.Web
         /// <summary>
         /// Initializes a new instance of <see cref="UmbracoHelper"/>.
         /// </summary>
-        /// <param name="umbracoContext">An Umbraco context.</param>
+        /// <param name="currentPage">The <see cref="IPublishedContent"/> item assigned to the helper.</param>
         /// <param name="tagQuery"></param>
         /// <param name="cultureDictionary"></param>
         /// <param name="componentRenderer"></param>
         /// <param name="publishedContentQuery"></param>
         /// <param name="membershipHelper"></param>
         /// <remarks>Sets the current page to the context's published content request's content item.</remarks>
-        public UmbracoHelper(UmbracoContext umbracoContext, ITagQuery tagQuery,
+        public UmbracoHelper(IPublishedContent currentPage,
+            ITagQuery tagQuery,
             ICultureDictionaryFactory cultureDictionary,
             IUmbracoComponentRenderer componentRenderer,
             IPublishedContentQuery publishedContentQuery,
             MembershipHelper membershipHelper)
         {
-            _umbracoContext = umbracoContext ?? throw new ArgumentNullException(nameof(umbracoContext));
             _tagQuery = tagQuery ?? throw new ArgumentNullException(nameof(tagQuery));
             _cultureDictionaryFactory = cultureDictionary ?? throw new ArgumentNullException(nameof(cultureDictionary));
             _componentRenderer = componentRenderer ?? throw new ArgumentNullException(nameof(componentRenderer));
             _membershipHelper = membershipHelper ?? throw new ArgumentNullException(nameof(membershipHelper));
             _publishedContentQuery = publishedContentQuery ?? throw new ArgumentNullException(nameof(publishedContentQuery));
-
-            if (_umbracoContext.IsFrontEndUmbracoRequest)
-                _currentPage = _umbracoContext.PublishedRequest.PublishedContent;
+            _currentPage = currentPage;
         }
 
         /// <summary>
@@ -77,7 +72,6 @@ namespace Umbraco.Web
 
         private IUmbracoComponentRenderer ComponentRenderer => Ensure(_componentRenderer);
         private ICultureDictionaryFactory CultureDictionaryFactory => Ensure(_cultureDictionaryFactory);
-        private UmbracoContext UmbracoContext => Ensure(_umbracoContext);
 
         /// <summary>
         /// Gets the tag context.
@@ -95,25 +89,19 @@ namespace Umbraco.Web
         public MembershipHelper MembershipHelper => Ensure(_membershipHelper);
 
         /// <summary>
-        /// Gets the url provider.
-        /// </summary>
-        public UrlProvider UrlProvider => UmbracoContext.UrlProvider;
-
-        /// <summary>
         /// Gets (or sets) the current <see cref="IPublishedContent"/> item assigned to the UmbracoHelper.
         /// </summary>
         /// <remarks>
         /// <para>
         /// Note that this is the assigned IPublishedContent item to the
         /// UmbracoHelper, this is not necessarily the Current IPublishedContent
-        /// item being rendered. This IPublishedContent object is contextual to
-        /// the current UmbracoHelper instance.
+        /// item being rendered that is assigned to the UmbracoContext.
+        /// This IPublishedContent object is contextual to the current UmbracoHelper instance.
         /// </para>
         ///<para>
         /// In some cases accessing this property will throw an exception if
         /// there is not IPublishedContent assigned to the Helper this will
-        /// only ever happen if the Helper is constructed with an UmbracoContext
-        /// and it is not a front-end request.
+        /// only ever happen if the Helper is constructed via DI during a non front-end request.
         /// </para>
         /// </remarks>
         /// <exception cref="InvalidOperationException">Thrown if the
@@ -129,7 +117,7 @@ namespace Umbraco.Web
                 }
 
                 throw new InvalidOperationException(
-                    $"Cannot return the {nameof(IPublishedContent)} because the {nameof(UmbracoHelper)} was constructed with an {nameof(UmbracoContext)} and the current request is not a front-end request."
+                    $"Cannot return the {nameof(IPublishedContent)} because the {nameof(UmbracoHelper)} was not constructed with an {nameof(IPublishedContent)}."
                     );
 
             }
@@ -156,7 +144,7 @@ namespace Umbraco.Web
         /// <returns></returns>
         public IHtmlString RenderMacro(string alias)
         {
-            return ComponentRenderer.RenderMacro(UmbracoContext.PublishedRequest?.PublishedContent?.Id ?? 0, alias, new { });
+            return ComponentRenderer.RenderMacro(AssignedContentItem?.Id ?? 0, alias, new { });
         }
 
         /// <summary>
@@ -167,7 +155,7 @@ namespace Umbraco.Web
         /// <returns></returns>
         public IHtmlString RenderMacro(string alias, object parameters)
         {
-            return ComponentRenderer.RenderMacro(UmbracoContext.PublishedRequest?.PublishedContent?.Id ?? 0, alias, parameters.ToDictionary<object>());
+            return ComponentRenderer.RenderMacro(AssignedContentItem?.Id ?? 0, alias, parameters.ToDictionary<object>());
         }
 
         /// <summary>
@@ -178,7 +166,7 @@ namespace Umbraco.Web
         /// <returns></returns>
         public IHtmlString RenderMacro(string alias, IDictionary<string, object> parameters)
         {
-            return ComponentRenderer.RenderMacro(UmbracoContext.PublishedRequest?.PublishedContent?.Id ?? 0, alias, parameters);
+            return ComponentRenderer.RenderMacro(AssignedContentItem?.Id ?? 0, alias, parameters);
         }
 
         #endregion
@@ -242,41 +230,7 @@ namespace Umbraco.Web
 
         #endregion
 
-        #region Urls
-
-        /// <summary>
-        /// Gets the url of a content identified by its identifier.
-        /// </summary>
-        /// <param name="contentId">The content identifier.</param>
-        /// <param name="culture"></param>
-        /// <returns>The url for the content.</returns>
-        public string Url(int contentId, string culture = null)
-        {
-            return UrlProvider.GetUrl(contentId, culture);
-        }
-
-        /// <summary>
-        /// Gets the url of a content identified by its identifier, in a specified mode.
-        /// </summary>
-        /// <param name="contentId">The content identifier.</param>
-        /// <param name="mode">The mode.</param>
-        /// <returns>The url for the content.</returns>
-        public string Url(int contentId, UrlProviderMode mode, string culture = null)
-        {
-            return UrlProvider.GetUrl(contentId, mode, culture);
-        }
-
-        /// <summary>
-        /// Gets the absolute url of a content identified by its identifier.
-        /// </summary>
-        /// <param name="contentId">The content identifier.</param>
-        /// <returns>The absolute url for the content.</returns>
-        public string UrlAbsolute(int contentId, string culture = null)
-        {
-            return UrlProvider.GetUrl(contentId, true, culture);
-        }
-
-        #endregion
+       
 
         #region Member/Content/Media from Udi
 
@@ -720,218 +674,10 @@ namespace Umbraco.Web
 
         #endregion
 
-        #region Strings
+       
 
-        /// <summary>
-        /// Replaces text line breaks with HTML line breaks
-        /// </summary>
-        /// <param name="text">The text.</param>
-        /// <returns>The text with text line breaks replaced with HTML line breaks (<br/>)</returns>
-        public IHtmlString ReplaceLineBreaksForHtml(string text)
-        {
-            return StringUtilities.ReplaceLineBreaksForHtml(text);
-        }
+        
 
-        /// <summary>
-        /// Generates a hash based on the text string passed in.  This method will detect the
-        /// security requirements (is FIPS enabled) and return an appropriate hash.
-        /// </summary>
-        /// <param name="text">The text to create a hash from</param>
-        /// <returns>Hash of the text string</returns>
-        public string CreateHash(string text)
-        {
-            return text.GenerateHash();
-        }
-
-        /// <summary>
-        /// Strips all HTML tags from a given string, all contents of the tags will remain.
-        /// </summary>
-        public HtmlString StripHtml(IHtmlString html, params string[] tags)
-        {
-            return StripHtml(html.ToHtmlString(), tags);
-        }
-
-        /// <summary>
-        /// Strips all HTML tags from a given string, all contents of the tags will remain.
-        /// </summary>
-        public HtmlString StripHtml(string html, params string[] tags)
-        {
-            return StringUtilities.StripHtmlTags(html, tags);
-        }
-
-        /// <summary>
-        /// Will take the first non-null value in the collection and return the value of it.
-        /// </summary>
-        public string Coalesce(params object[] args)
-        {
-            return StringUtilities.Coalesce(args);
-        }
-
-        /// <summary>
-        /// Joins any number of int/string/objects into one string
-        /// </summary>
-        public string Concatenate(params object[] args)
-        {
-            return StringUtilities.Concatenate(args);
-        }
-
-        /// <summary>
-        /// Joins any number of int/string/objects into one string and separates them with the string separator parameter.
-        /// </summary>
-        public string Join(string separator, params object[] args)
-        {
-            return StringUtilities.Join(separator, args);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given length, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString Truncate(IHtmlString html, int length)
-        {
-            return Truncate(html.ToHtmlString(), length, true, false);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given length, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString Truncate(IHtmlString html, int length, bool addElipsis)
-        {
-            return Truncate(html.ToHtmlString(), length, addElipsis, false);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given length, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString Truncate(IHtmlString html, int length, bool addElipsis, bool treatTagsAsContent)
-        {
-            return Truncate(html.ToHtmlString(), length, addElipsis, treatTagsAsContent);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given length, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString Truncate(string html, int length)
-        {
-            return Truncate(html, length, true, false);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given length, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString Truncate(string html, int length, bool addElipsis)
-        {
-            return Truncate(html, length, addElipsis, false);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given length, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString Truncate(string html, int length, bool addElipsis, bool treatTagsAsContent)
-        {
-            return StringUtilities.Truncate(html, length, addElipsis, treatTagsAsContent);
-        }
-
-        #region Truncate by Words
-
-        /// <summary>
-        /// Truncates a string to a given amount of words, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString TruncateByWords(string html, int words)
-        {
-            int length = StringUtilities.WordsToLength(html, words);
-
-            return Truncate(html, length, true, false);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given amount of words, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString TruncateByWords(string html, int words, bool addElipsis)
-        {
-            int length = StringUtilities.WordsToLength(html, words);
-
-            return Truncate(html, length, addElipsis, false);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given amount of words, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString TruncateByWords(IHtmlString html, int words)
-        {
-            int length = StringUtilities.WordsToLength(html.ToHtmlString(), words);
-
-            return Truncate(html, length, true, false);
-        }
-
-        /// <summary>
-        /// Truncates a string to a given amount of words, can add a ellipsis at the end (...). Method checks for open HTML tags, and makes sure to close them
-        /// </summary>
-        public IHtmlString TruncateByWords(IHtmlString html, int words, bool addElipsis)
-        {
-            int length = StringUtilities.WordsToLength(html.ToHtmlString(), words);
-
-            return Truncate(html, length, addElipsis, false);
-        }
-
-        #endregion
-
-        #endregion
-
-        #region If
-
-        /// <summary>
-        /// If the test is true, the string valueIfTrue will be returned, otherwise the valueIfFalse will be returned.
-        /// </summary>
-        public HtmlString If(bool test, string valueIfTrue, string valueIfFalse)
-        {
-            return test ? new HtmlString(valueIfTrue) : new HtmlString(valueIfFalse);
-        }
-
-        /// <summary>
-        /// If the test is true, the string valueIfTrue will be returned, otherwise the valueIfFalse will be returned.
-        /// </summary>
-        public HtmlString If(bool test, string valueIfTrue)
-        {
-            return test ? new HtmlString(valueIfTrue) : new HtmlString(string.Empty);
-        }
-
-        #endregion
-
-        /// <summary>
-        /// This is used in methods like BeginUmbracoForm and SurfaceAction to generate an encrypted string which gets submitted in a request for which
-        /// Umbraco can decrypt during the routing process in order to delegate the request to a specific MVC Controller.
-        /// </summary>
-        /// <param name="controllerName"></param>
-        /// <param name="controllerAction"></param>
-        /// <param name="area"></param>
-        /// <param name="additionalRouteVals"></param>
-        /// <returns></returns>
-        internal static string CreateEncryptedRouteString(string controllerName, string controllerAction, string area, object additionalRouteVals = null)
-        {
-            if (string.IsNullOrEmpty(controllerName)) throw new ArgumentNullOrEmptyException(nameof(controllerName));
-            if (string.IsNullOrEmpty(controllerAction)) throw new ArgumentNullOrEmptyException(nameof(controllerAction));
-            if (area == null) throw new ArgumentNullException(nameof(area));
-
-            //need to create a params string as Base64 to put into our hidden field to use during the routes
-            var surfaceRouteParams = $"c={HttpUtility.UrlEncode(controllerName)}&a={HttpUtility.UrlEncode(controllerAction)}&ar={area}";
-
-            //checking if the additional route values is already a dictionary and convert to querystring
-            string additionalRouteValsAsQuery;
-            if (additionalRouteVals != null)
-            {
-                var additionalRouteValsAsDictionary = additionalRouteVals as Dictionary<string, object>;
-                if (additionalRouteValsAsDictionary != null)
-                    additionalRouteValsAsQuery = additionalRouteValsAsDictionary.ToQueryString();
-                else
-                    additionalRouteValsAsQuery = additionalRouteVals.ToDictionary<object>().ToQueryString();
-            }
-            else
-                additionalRouteValsAsQuery = null;
-
-            if (additionalRouteValsAsQuery.IsNullOrWhiteSpace() == false)
-                surfaceRouteParams += "&" + additionalRouteValsAsQuery;
-
-            return surfaceRouteParams.EncryptWithMachineKey();
-        }
+        
     }
 }
